feat: handle error if file already exists

+++ src/fs/create.js
@@ -1,9 +1,12 @@
+import { existsSync } from 'node:fs';
 import { writeFile } from 'node:fs/promises';
 
 const create = async () => {
     const path = './files/fresh.txt';
-    const content = 'I am fresh and young'
 
+    if (existsSync(path)) throw new Error('FS operation failed');
+
+    const content = 'I am fresh and young';
     writeFile(path, content);
 };
